requirements.txt
requests
pyyaml
schedule
tweepy
linkedin-api
spacy
streamlit
pandas
supabase
